Wired up db client and started gql schemas/resolvers

+++ packages/pfa-influxr/src/sql/02-reporting/03-tables/schedule.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS reporting.schedule (
-    id_game     integer,
+    id_game     bigint,
     season      integer,
     week        integer,
     day         varchar(3),
